Update plot.py
Added the option of 'animation' to the plot. Please review it. The animation can be interacted with by clicking the matplotlib widgets. The modification doesn't effect anything else in the plot. I also tried to create a Video object, but that was a disaster. It was horribly slow. Let's discuss this modification. I think it would be great for other purposes as this is a simple way of visualizing video data without cv2.