Update plot.py
Modified the previous version in order to have a PlotAnimation object for cleaner code and better handling of passing variables.
The speed up and slow down button is changed to frames per second setting button. The toggle button is changed to start/pause button.
Known bugs: Reset and speed setting is only working only when the mouse is moved over from the top of the button, which can be a little annoying sometimes.
Place for improvement: EFIT overlay.