Update plot.py
This time I am not going to do a pull request (as my last one was kind of ignored after I fixed a bug in it). This is the development package. For me the animation is working with the GPI data. It should be tested with other animation. I also added a sample configuration file for the EFIT overlay. It is written for NSTX.
I tested the animation with the flap_tests animation test fraction and it was generating the video and showing it.
Otherwise, other parts of the code are not affected by this modification.