Update data_object.py
Added the time property for the data object. It is not a set property, it doesn't increase the memory usage, it is just returning the DataObject.coordinate('Time')[0] for convenience. The corresponding code needs to be checked. It is quite simple, but the notation may not be general enough. There is no name mangling or anything like that.
Also, I added the data_source=data_source instead of the data_source=None expression in the def get_data() method.