Video saving bugfix and tools.py enhancement
plot.py:
cv2 video saving was in RGBA but for cv2 it needs to be BGR to have correct colors in the avi file.
tools.py:
time_unit_translation was modified so it could translate the time from the maximum value of the time if no unit is given. This could occur in MDSplus and then the plot won't have a correct time axis.
Backwards time unit translation is also supported: e.g. 1e-3 --> ms.