Various bugfixes
data_object: keyerror is raised because that error handling is for the data_source input keyerror and not a typeerror which might occur within.
plot.py: Compacted the code whereever I found reused lines, added the exp_id to the anim-image and anim-contour. Added a memo for a bugfix in the future.
tools.py: There was an error in handling the type of the time unit. It could have been either string or a  numpy string and that caused error during mdsplus data read.