Different unit based plotting
plot.py and coordinate.py were modified for different unit plotting. It works as the following:
plot.py can accept in the options a following option (just an example):
{'Plot units':{'Time':'ms','Device R': 'm', 'Device z': 'm'}}

The plotted data is converted to the appropriate unit.

Only the animation part is modified, I didn't want to touch the others, yet.

Known limitation: the EFIT overplotting is not working with this option yet.