Plot animation fix and spectral_analysis bugfix
plot.py: rearranged some lines and fixed the difference between anim-contour and anim-image. There is still some bug in the pdd_list and the coordinate indices, but I couldn't fix it without significantly rewriting codes which are already operational.
spectral_analysis.py:
Trend removal failed if used on int16 data. If I modifiy the code to this one it works and the functionality is the same.