Path and contour overplotting
This revision is a general overplotting option for the plot.py. It has replaced the 'EFIT options' as that has only been used by me.
It can handle any number of contours and paths. It needs to be in the same time unit as the original data and in the format like the default_overplot_options variable.