Overplotting, equal axes and axes visibility options
Overplotting options extended, they are working now for all the 2D plot types. Also added the option of 'Equal axes' and 'Axes visibility'.
The former maintains equal x and y axis if their coordinate units are the same. The latter removes the xtitle and the ytitle from the axes. It's useful when plotting multiple plots.