Spectral analysis 2D temporary solution and plot bugfix
Spectral analysis couldn't do anything about a 2D CCF function, it was failing all the time. A temporary solution has been implemented, which only work with 2D data, but not with 2+1D. A solution needs to be found, but it serves the purpuse. It is also a bit slow compared to the scipy CCF calculator.